Added some fun changes for players
Removed the tracer from being visible from the player.

Added explosions and allowed it to be applied to breakable objects that the player can interact with, as well as making the enemy blow up upon defeat in order to reduce entities. (Also the explosion is based on Power Ranger enemies blowing up on defeat). Also added blood FX upon hitting the enemy.

(Added some VFX files and creature files too)